Invalid input now prevents answer from being shown.
Had to change script directory path to include /solu-buddy/ directory. Not sure why this is required or how to fix it.

TODO: Need to prevent is_valid_formula(str) from treating "Nacl" as valid. This can probably be done with a check comparing number of lowercase to uppercase. The number of lowercase must always be less than or equal to number of uppercase